modify: where requested change POST to initially requested GET method, rename "payload" to "api" package
