+++ src/main/java/com/epam/labtaskspringcore/api/TraineeUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.epam.labtaskspringcore.payloads;
+package com.epam.labtaskspringcore.api;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
