+++ src/main/java/com/epam/labtaskspringcore/api/TrainerDTOWithTraineeList.java
@@ -1,34 +1,34 @@
-package com.epam.labtaskspringcore.payloads;
+package com.epam.labtaskspringcore.api;
 
 import com.epam.labtaskspringcore.model.TrainingType;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 
-@Getter
-@Setter
-public class TrainerUpdateRequest extends UsernamePassword {
+import java.util.Set;
 
-    @NotBlank
+@Setter
+@Getter
+public class TrainerDTOWithTraineeList {
+
     private String firstName;
-    @NotBlank
     private String lastName;
     private TrainingType specialization;
-    @NotNull
     private Boolean isActive;
+    Set<TraineeDTOForTraineesList> trainees;
 
-    public TrainerUpdateRequest(
-            String username,
-            String password,
+    public TrainerDTOWithTraineeList(
             String firstName,
             String lastName,
             TrainingType specialization,
-            Boolean isActive) {
-        super(username, password);
+            Boolean isActive,
+            Set<TraineeDTOForTraineesList> trainees) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.specialization = specialization;
         this.isActive = isActive;
+        this.trainees = trainees;
+    }
+
+    public TrainerDTOWithTraineeList() {
     }
 }
