+++ src/main/java/com/epam/labtaskspringcore/api/TrainerUpdateRequest.java
@@ -1,34 +1,34 @@
-package com.epam.labtaskspringcore.payloads;
+package com.epam.labtaskspringcore.api;
 
 import com.epam.labtaskspringcore.model.TrainingType;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.util.Set;
-
-@Setter
 @Getter
-public class TrainerDTOWithTraineeList {
+@Setter
+public class TrainerUpdateRequest extends UsernamePassword {
 
+    @NotBlank
     private String firstName;
+    @NotBlank
     private String lastName;
     private TrainingType specialization;
+    @NotNull
     private Boolean isActive;
-    Set<TraineeDTOForTraineesList> trainees;
 
-    public TrainerDTOWithTraineeList(
+    public TrainerUpdateRequest(
+            String username,
+            String password,
             String firstName,
             String lastName,
             TrainingType specialization,
-            Boolean isActive,
-            Set<TraineeDTOForTraineesList> trainees) {
+            Boolean isActive) {
+        super(username, password);
         this.firstName = firstName;
         this.lastName = lastName;
         this.specialization = specialization;
         this.isActive = isActive;
-        this.trainees = trainees;
-    }
-
-    public TrainerDTOWithTraineeList() {
     }
 }
