+++ src/main/java/com/epam/labtaskspringcore/controller/UserManagementController.java
@@ -2,8 +2,8 @@ package com.epam.labtaskspringcore.controller;
 
 import com.epam.labtaskspringcore.aspect.CheckUsernamePassword;
 import com.epam.labtaskspringcore.aspect.LogRestDetails;
-import com.epam.labtaskspringcore.payloads.PasswordUpdateRequest;
-import com.epam.labtaskspringcore.payloads.UsernamePassword;
+import com.epam.labtaskspringcore.api.PasswordUpdateRequest;
+import com.epam.labtaskspringcore.api.UsernamePassword;
 import com.epam.labtaskspringcore.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -13,10 +13,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @Slf4j
 @RestController
@@ -35,9 +32,8 @@ public class UserManagementController {
     @ApiResponses(value = {
             @ApiResponse(responseCode = "204", description = "No Content")
     })
-    @PostMapping("/user/login")
+    @GetMapping("/user/login")
     public ResponseEntity<?> login(@Valid @RequestBody UsernamePassword usernamePassword) {
-        // response code 204 because there is no content
         return ResponseEntity.status(HttpStatusCode.valueOf(204)).build();
     }
 
@@ -47,14 +43,10 @@ public class UserManagementController {
     })
     @PutMapping("/user/login")
     public ResponseEntity<?> updatePassword(@Valid @RequestBody PasswordUpdateRequest usernamePassword) {
-
         String username = usernamePassword.getUsername();
         String currentPassword = usernamePassword.getPassword();
         String newPassword = usernamePassword.getNewPassword();
-
         userService.updatePassword(username, currentPassword, newPassword);
-
-        // response code 204 because there is no content
         return ResponseEntity.status(HttpStatusCode.valueOf(204)).build();
     }
 }
